successfully migrated application from postgree to maria database

+++ Jokes/src/main/java/com/jokes/vicevi/entities/Category.java
@@ -13,7 +13,7 @@ public class Category {
 @Id
 @GeneratedValue(strategy=GenerationType.IDENTITY)
 @Column(name="id")
-private Long id;
+private Integer id;
 
 @NotBlank
 @Column(unique=true)
@@ -24,19 +24,19 @@ private List<Jokes> vicevi;
 
 public Category() {}
 
-public Category(Long id,String name) {
+public Category(Integer id,String name) {
 	
     this.id=id;
 	this.name = name;
 }
 
 
-public Long getId() {
+public Integer getId() {
 	return id;
 }
 
 
-public void setId(Long id) {
+public void setId(Integer id) {
 	this.id = id;
 }
